v1.2
Added direct image downloading support using argument "images:<url>". Since this is just a single file, progress isn't important but it reports when it's finished (if enabled) and it uses a class file over a form. The protocol for this is optional.
Changed protocol installation procedure. You can now specify if you want to install the tags or pools protocol instead of both at the same time. The images protocol is optional and won't trigger the "Protocol" menu option.
I added a technical form, but it isn't gonna be expanded upon, I original intended to use it for queuing information for pools + images but I never got around to working on it since my original vision for it got faded. If I figure it out again, I may add it for people who just want the information instead of pictures.
New images userscript for downloading from the site. Click a button near the tag bar.

its gonna be (alright)