v1.4
Added 3 new forms: Pool Wishlist, Redownloader, and Indexer.
Pool Wishlist is basically a storage for pools to download later, such as incomplete or in-progress pools.
Redownloader compiles a list of downloaded pools and tags which allow you to redownload from the application. I suggest manually redownloading the tags before it will be useful as there is a new addition to the tag downloader which will make this an easy to use tool.
Indexer is a form that tests the e621 api and returns information, not all the information but most. The only missing info it doesn't return is if it's blacklisted or if it'll be skipped or not downloaded.

Changed tag downloading, it saves all the tags in tags.nfo & tags.blacklisted.nfo. Used for the redownloader, so redownload tags OR add a line at the very top of both of the files with "TAGS: " and the tags you downloaded, example: "TAGS: chunie gay".
Removed the technical form, replaced with indexer.
Bug fixes, if any.