fixed some code and added new custom controls
renamed some as well, i might have to clean up the github again lol

Majorly: some options are changed again. Sorry, I'm trying not to mess with them, but at this point I might make them more identifiable.

but basically, there's a button inside a text box with the option for hints (extended textbox), and a button with a UAC shield (extended button).

pools are now one-loop for all API pages! I like how I handled the files this way, it's easier to manage and less if-else if-else if, by having the full paths in one list, so i just need one webclient to download. I might do the same for tags, but only have 3 loops for clean, graylisted, and blacklisted.

images are also more one-class controlled!

and tags, well they're tags.