finally got new tag download working
i learned that BeginInvoke invokes the controls while the thread is running, which could make outofbounds exceptions occur if it tries to call an index that could be different when the loop ends. i switched it to Invoke instead, which apparently invokes the form AND THEN goes to download, instead of possibly downloading while invoking. I applied this new knowledge to the other downloaders. Praise be to tom hanks.

i also updated a few controls to clean up Code Analysis. There's only 1 warning, and quite frankly it's not my problem to fix it.